Fix redundant casts/conditions in psalm (#1874)
This includes a check for array_map that was accidentally an assignment

+++ src/Psalm/Internal/Codebase/Reflection.php
@@ -160,7 +160,7 @@ class Reflection
         }
 
         $reflection_methods = $reflected_class->getMethods(
-            (int) (\ReflectionMethod::IS_PUBLIC | \ReflectionMethod::IS_PROTECTED)
+            (\ReflectionMethod::IS_PUBLIC | \ReflectionMethod::IS_PROTECTED)
         );
 
         if ($class_name_lower === 'generator') {
